+++ cloudtower-java-sdk-okhttp-gson/src/main/java/com/smartx/tower/model/BrickTopoCapacity.java
@@ -32,14 +32,14 @@ import org.openapitools.jackson.nullable.JsonNullable;
 public class BrickTopoCapacity {
   public static final String SERIALIZED_NAME_ROW = "row";
   @SerializedName(SERIALIZED_NAME_ROW)
-  private Double row;
+  private Integer row;
 
   public static final String SERIALIZED_NAME_COLUMN = "column";
   @SerializedName(SERIALIZED_NAME_COLUMN)
-  private Double column;
+  private Integer column;
 
 
-  public BrickTopoCapacity row(Double row) {
+  public BrickTopoCapacity row(Integer row) {
     
     this.row = row;
     return this;
@@ -52,17 +52,17 @@ public class BrickTopoCapacity {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Double getRow() {
+  public Integer getRow() {
     return row;
   }
 
 
-  public void setRow(Double row) {
+  public void setRow(Integer row) {
     this.row = row;
   }
 
 
-  public BrickTopoCapacity column(Double column) {
+  public BrickTopoCapacity column(Integer column) {
     
     this.column = column;
     return this;
@@ -75,12 +75,12 @@ public class BrickTopoCapacity {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Double getColumn() {
+  public Integer getColumn() {
     return column;
   }
 
 
-  public void setColumn(Double column) {
+  public void setColumn(Integer column) {
     this.column = column;
   }
